Move backend and frontend files to new structure
Migrated backend source files from src/tokopaedi to backend/src/tokopaedi and added backend/src/main.py for FastAPI endpoints. Added initial frontend project files including configuration, assets, and app source code to the frontend directory. This refactor separates backend and frontend codebases for improved project organization.
